requirements.txt
aiogram==2.25.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2023.3
redis==5.0.7
SQLAlchemy==1.4.54
asyncpg==0.29.0